[NUGUDI-66] refactor(web): 컴포넌트 경로 수정
