+++ apps/web/app/auth/sign-in/email/page.tsx
@@ -1,4 +1,4 @@
-import EmailLoginView from "@/app/_ui/views/email-login-view/email-login-view";
+import EmailLoginView from "@/app/_ui/views/email-login-view";
 
 const EmailLoginPage = () => {
   return <EmailLoginView />;
